Выносит renderChart.js из statistics-component.js
